some systems may not like using a size_t for a %d in a format, use an int.
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@16679 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/BanCommands.cxx
@@ -1093,7 +1093,7 @@ bool MasterBanCommand::operator() (const char	 *message,
     if (bans.size() > 20) {
       sendMessage(ServerPlayer, t,
 		  TextUtils::format("There are %d bans, only displaying the "
-				    "first 20", bans.size()).c_str());
+				    "first 20", (int)bans.size()).c_str());
 
     } else if (bans.size() == 0) {
       sendMessage(ServerPlayer, t, "There are no master bans loaded.");
